requirements.txt
Flask~=3.1.0
Flask-Cors~=5.0.1
llama-index-core>=0.11.0
# see https://llamahub.ai for all integrations
llama-index-embeddings-openai  # can replace with the embed model you want to use
llama-index-llms-openai  # can replace with the llm you want to use
llama-index-agent-openai

Werkzeug~=3.1.3